+++ CHANGELOG.md
@@ -14,6 +14,7 @@ DS
 UI
 - Text input via IME was broken
 - Double clicking in text field may lose focus
+- Scroll deltas are consumed even the text or text field is not scrollable
 
 ## UI [2.0.1] - 2025-01-25